+++ src/main/kotlin/com/calmwolfs/bedwar/features/notifications/GameNotifications.kt
@@ -5,7 +5,6 @@ import com.calmwolfs.bedwar.events.bedwars.BedBreakEvent
 import com.calmwolfs.bedwar.events.bedwars.FinalKillEvent
 import com.calmwolfs.bedwar.events.bedwars.KillEvent
 import com.calmwolfs.bedwar.features.team.TeamStatus
-import com.calmwolfs.bedwar.utils.HypixelUtils
 import com.calmwolfs.bedwar.utils.gui.NotificationUtils
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
@@ -16,11 +15,11 @@ object GameNotifications {
     //todo make it say "You" instead of your username
     @SubscribeEvent
     fun onKill(event: KillEvent) {
-        if (gameConfig.enabled && event.killer in TeamStatus.currentTeamMembers || event.killer == HypixelUtils.currentName) {
+        if (gameConfig.enabled && event.killer in TeamStatus.currentTeamMembers) {
             val message = formatNotification(listOf("§3${event.killer} §agot a kill!"))
             NotificationUtils.displayNotification(message, null, gameConfig.sound, gameConfig.displayLength.toDouble())
         }
-        if (deathConfig.enabled && event.killed in TeamStatus.currentTeamMembers || event.killed == HypixelUtils.currentName) {
+        if (deathConfig.enabled && event.killed in TeamStatus.currentTeamMembers) {
             val message = formatNotification(listOf("§3${event.killed} §ewas killed!"))
             NotificationUtils.displayNotification(message, null, deathConfig.sound, deathConfig.displayLength.toDouble())
         }
@@ -28,11 +27,11 @@ object GameNotifications {
 
     @SubscribeEvent
     fun onFinal(event: FinalKillEvent) {
-        if (gameConfig.enabled && event.killer in TeamStatus.currentTeamMembers || event.killer == HypixelUtils.currentName) {
+        if (gameConfig.enabled && event.killer in TeamStatus.currentTeamMembers) {
             val message = formatNotification(listOf("§3${event.killer} §agot a final kill!"))
             NotificationUtils.displayNotification(message, null, gameConfig.sound, gameConfig.displayLength.toDouble())
         }
-        if (deathConfig.enabled && event.killed in TeamStatus.currentTeamMembers || event.killed == HypixelUtils.currentName) {
+        if (deathConfig.enabled && event.killed in TeamStatus.currentTeamMembers) {
             val message = formatNotification(listOf("§3${event.killed} §ewas final killed!"))
             NotificationUtils.displayNotification(message, null, deathConfig.sound, deathConfig.displayLength.toDouble())
         }
@@ -40,8 +39,7 @@ object GameNotifications {
 
     @SubscribeEvent
     fun onBed(event: BedBreakEvent) {
-        if (!gameConfig.enabled) return
-        if (event.player in TeamStatus.currentTeamMembers || event.player == HypixelUtils.currentName) {
+        if (gameConfig.enabled && event.player in TeamStatus.currentTeamMembers) {
             val message = formatNotification(listOf("§3${event.player} §abroke a bed!"))
             NotificationUtils.displayNotification(message, null, gameConfig.sound, gameConfig.displayLength.toDouble())
         }